@@ -3,7 +3,7 @@ require 'spec_helper'
 describe RestPack::Service::Configuration do
   subject { RestPack::Service::Configuration.new }
 
-  it "has defaults" do
-    subject.application_id.should == 1
+  it 'has defaults' do
+    expect(subject.application_id).to eq(1)
   end
 end
